refactor: files to be where it should be
