+++ src/components/Button.jsx
@@ -1,49 +1,18 @@
 const Button = ({ value }) => {
-  let style =
-    'h-12 w-1/5 rounded bg-themeOne-80 text-3xl text-themeOne-100 shadow-ash hover:bg-themeOne-90 active:translate-y-1';
-
-  switch (value) {
-    case 'DEL':
-      style =
-        'delete-btn h-12 w-1/5 rounded bg-themeOne-120 text-2xl font-bold text-themeOne-white shadow-darkblue hover:bg-themeOne-130 active:translate-y-1';
-      break;
-    case 'RESET':
-      style =
-        ' h-12 w-[48%] rounded bg-themeOne-150 text-2xl font-bold text-themeOne-white shadow-darkblue hover:bg-themeOne-160 active:translate-y-1';
-      break;
-    case '=':
-      style =
-        ' h-12 w-[48%] rounded bg-themeOne-180 text-2xl font-bold text-themeOne-white shadow-red hover:bg-themeOne-190 active:translate-y-1';
-      break;
-    default:
-      break;
-  }
+  const style =
+    value === 'DEL'
+      ? 'w-1/5 bg-themeOne-120 text-2xl font-bold text-themeOne-white shadow-darkblue hover:bg-themeOne-130'
+      : value === 'RESET'
+        ? 'w-[48%] bg-themeOne-150 text-2xl font-bold text-themeOne-white shadow-darkblue hover:bg-themeOne-160'
+        : value === '='
+          ? 'w-[48%] bg-themeOne-180 text-2xl font-bold text-themeOne-white shadow-red hover:bg-themeOne-190'
+          : 'w-1/5 bg-themeOne-80 text-3xl text-themeOne-100 shadow-ash hover:bg-themeOne-90';
 
   return (
-    <button className={style} type="button" value={value}>
+    <button className={`${style} h-12 rounded active:translate-y-1`} type="button" value={value}>
       {value}
     </button>
   );
 };
 
-const CalcButtons = () => {
-  const ButtonValues = [
-    ['7', '8', '9', 'DEL'],
-    ['4', '5', '6', '+'],
-    ['1', '2', '3', '-'],
-    ['.', '0', '/', 'x'],
-    ['RESET', '='],
-  ];
-
-  const MapButton = ButtonValues.map((row) => (
-    <div className="mb-2 flex items-center justify-between">
-      {row.map((buttonValue) => (
-        <Button value={buttonValue} />
-      ))}
-    </div>
-  ));
-
-  return <>{MapButton}</>;
-};
-
-export default CalcButtons;
+export default Button;
