+++ src/components/ButtonContainer.jsx
@@ -1,4 +1,24 @@
-import CalcButtons from './Button';
+import Button from './Button';
+
+const CalcButtons = () => {
+  const buttonValue = [
+    ['7', '8', '9', 'DEL'],
+    ['4', '5', '6', '+'],
+    ['1', '2', '3', '-'],
+    ['.', '0', '/', 'x'],
+    ['RESET', '='],
+  ];
+
+  const MapButton = buttonValue.map((row) => (
+    <div className="mb-2 flex items-center justify-between">
+      {row.map((buttonValue) => (
+        <Button value={buttonValue} />
+      ))}
+    </div>
+  ));
+
+  return <>{MapButton}</>;
+};
 
 const ButtonContainer = () => {
   return (
