+++ src/components/Input.jsx
@@ -1,7 +1,7 @@
 const Input = ({ id, value, checked, className }) => {
   return (
     <input
-      className={`'duration-200 h-4 w-4 cursor-pointer appearance-none rounded-full transition-all'${className}`}
+      className={`h-4 w-4 cursor-pointer appearance-none rounded-full transition-all duration-200 ${className}`}
       id={id}
       name="theme"
       type="radio"
