+++ src/components/Label.jsx
@@ -1,9 +1,3 @@
-const Themes = [
-  { id: 'theme-choice-one', value: '1' },
-  { id: 'theme-choice-two', value: '2' },
-  { id: 'theme-choice-three', value: '3' },
-];
-
 const Label = ({ value, htmlFor }) => {
   return (
     <label className="flex w-4 justify-center" htmlFor={htmlFor}>
@@ -12,4 +6,4 @@ const Label = ({ value, htmlFor }) => {
   );
 };
 
-export { Themes, Label };
+export default Label;
